backend/login-signup/python_scripts/requirements.txt
cerebras-cloud-sdk>=0.3.2
opencv-python-headless>=4.9.0.80
pytesseract>=0.3.10
Pillow>=10.3.0
numpy>=1.26.0
PyMuPDF>=1.24.10
easyocr>=1.7.2
torch>=2.2.0
torchvision>=0.17.0
